refactor: move into subpackage
